# environment-a/api-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BreadcrumbResponse:
    properties:
      breadcrumb:
        items:
          $ref: '#/definitions/dto.DirectoryDTO'
        type: array
      directory_id:
        type: string
      status:
        type: string
    type: object
  dto.CreateDirectoryRequest:
    properties:
      name:
        maxLength: 255
        type: string
      parent_id:
        type: string
    required:
    - name
    type: object
  dto.DirectoryDTO:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      depth:
        type: integer
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      starred_at:
        type: string
    type: object
  dto.DirectoryMutationResponse:
    properties:
      directory:
        $ref: '#/definitions/dto.DirectoryDTO'
      status:
        type: string
    type: object
  dto.DirectoryResponse:
    properties:
      directory:
        $ref: '#/definitions/dto.DirectoryDTO'
      status:
        type: string
    type: object
  dto.DirectoryTreeResponse:
    properties:
      directories:
        items:
          $ref: '#/definitions/dto.DirectoryDTO'
        type: array
      root_id:
        type: string
      status:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: unauthorized
        type: string
      status:
        example: error
        type: string
    type: object
  dto.FileDTO:
    properties:
      chunk_count:
        type: integer
      created_at:
        type: string
      dedup_ratio:
        type: number
      deleted_at:
        type: string
      directory_id:
        type: string
      id:
        type: string
      manifest_id:
        type: string
      mime_type:
        type: string
      name:
        type: string
      new_chunk_count:
        type: integer
      reuse_chunk_count:
        type: integer
      size_bytes:
        type: integer
      starred_at:
        type: string
    type: object
...
  dto.FileListResponse:
    properties:
      directory_id:
        type: string
      files:
        items:
          $ref: '#/definitions/dto.FileDTO'
        type: array
      status:
        type: string
    type: object
...
  dto.OKResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
  dto.StarredResponse:
    properties:
      directories:
        items:
          $ref: '#/definitions/dto.DirectoryDTO'
        type: array
      files:
        items:
          $ref: '#/definitions/dto.FileDTO'
        type: array
      status:
        type: string
    type: object
  dto.TrashResponse:
    properties:
      directories:
        items:
          $ref: '#/definitions/dto.DirectoryDTO'
        type: array
      files:
        items:
          $ref: '#/definitions/dto.FileDTO'
        type: array
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: API service for authentication, directory hierarchy, and immutable
    file metadata.
  title: HashBox API Service
  version: "1.0"
paths:
...
  /directories:
    post:
      consumes:
      - application/json
      description: Create root or nested directory in user namespace
      parameters:
      - description: Create directory payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDirectoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DirectoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create directory
      tags:
      - directories
  /directories/{id}:
    delete:
      description: Soft-delete a directory subtree and move it to trash
      parameters:
      - description: Directory UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DirectoryMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Move directory to trash
      tags:
      - directories
  /directories/{id}/breadcrumb:
    get:
      description: Resolve breadcrumb path for given directory
      parameters:
      - description: Directory UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BreadcrumbResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Directory breadcrumb
      tags:
      - directories
  /directories/{id}/files:
    get:
      description: List files by directory with optional include_deleted=true
      parameters:
      - description: Directory UUID
        in: path
        name: id
        required: true
        type: string
      - description: Include soft-deleted files
        in: query
        name: include_deleted
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FileListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Directory files
      tags:
      - directories
  /directories/{id}/permanent:
    delete:
      description: Permanently remove a directory subtree from trash metadata
      parameters:
      - description: Directory UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Permanently delete directory
      tags:
      - directories
  /directories/{id}/restore:
    post:
      description: Restore a previously deleted directory subtree
      parameters:
      - description: Directory UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DirectoryMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore directory from trash
      tags:
      - directories
  /directories/{id}/star:
    delete:
      description: Remove starred marker from a directory
      parameters:
      - description: Directory UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DirectoryMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unstar directory
      tags:
      - directories
    put:
      description: Mark a directory as starred
      parameters:
      - description: Directory UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DirectoryMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Star directory
      tags:
      - directories
  /directories/tree:
    get:
      description: List root directories or subtree with root_id query
      parameters:
      - description: Root directory UUID
        in: query
        name: root_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DirectoryTreeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Directory tree
      tags:
      - directories
...
  /starred:
    get:
      description: List user's starred files and directories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StarredResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Starred items
      tags:
      - workspace
...
  /trash:
    get:
      description: List user's deleted files and directory trash roots
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrashResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Trash
      tags:
      - workspace
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"